requirements.txt
coverage==6.5.0
Django==4.1.3
djangorestframework==3.12.0
drf-yasg2==1.19.4
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
